Create newEl.js to extract createEl() from player.js. Leaves variables(tag, el, playerElIngest, divEmbed, attrs) in player.js, and if-statement include super.
Change every "this" to "player" in newEl.

Issue #2
